requirements.txt
boto3==1.28.79
botocore==1.31.79
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
idna==3.4
jmespath==1.0.1
kafka-python==2.0.2
mysql-connector-python==8.2.0
protobuf==4.21.12
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
urllib3==2.0.7